Change history-tab directive to component and test it to get 100%
