+++ core/templates/pages/exploration-editor-page/history-tab/history-tab.component.html
